Fix DataSizeFixer

+++ src/service/fixer/parse_error_fixer.rs
@@ -18,21 +18,21 @@ impl ParseErrorFixer {
         println!("Fixing a parse error...");
 
         // Check file size
+        // TODO: Support chained FIT files
         let file = fs::read(filename)?;
         let file_size = file.len();
         let fit = Fit::read(file)?;
 
-        let actual_data_size = fit.header.data_size as usize;
-        let expected_data_size = file_size - fit.header.header_size as usize;
+        let actual_data_size = fit.header.data_size;
+        let expected_data_size = file_size as u32 - fit.header.header_size as u32;
         if actual_data_size != expected_data_size {
             println!(
                 "A data size is broken. Expected: {}, Actual: {}",
                 expected_data_size, actual_data_size
             );
 
-            // TODO: Check expected_data_size must be within u32 range
             let data_size_fixer = DataSizeFixer::new();
-            data_size_fixer.fix(filename, expected_data_size as u32)?;
+            data_size_fixer.fix(filename, expected_data_size)?;
         }
         println!("A parse error has been fixed.");
